Correct motion vectors and frame handling
The global frame motion vector was not being decoded correctly (was not
taking sign into account). This was never caught because I guess nobody
ever tried to decode a file that used non-zero motion vectors.

This commit also adds a special hack for frame handling. RoQ assumes a
pair of buffers for decoding. Decoding skip blocks on frame 2 is supposed
to leave those blocks with the same data as frame 0, since it's the same
memory. However, frame 1 can still have skip blocks. Deal with this by
copying decoded frame 0 to frame 1 before proceeding with frame 1 decode.