Add audio file output to test application
Hook up the audio callback in the test application so that it outputs
the audio data to a standard WAV file. Also, add a "playback finished"
callback so that a client application can clean up anything it needs
to (in the case of the test app, it needs the signal in order to finish
writing the WAV file).